refactor: Implement MotionTrackerBD
