+++ alvr/client_openxr/src/extra_extensions/body_tracking_bd.rs
@@ -5,8 +5,6 @@ use std::ffi::{c_char, c_void, CString};
 use std::ptr;
 
 pub const BD_BODY_TRACKING_EXTENSION_NAME: &str = "XR_BD_body_tracking";
-pub const BD_MOTION_TRACKING_EXTENSION_NAME: &str = "XR_BD_motion_tracking";
-pub const PICO_CONFIGURATION_EXTENSION_NAME: &str = "XR_PICO_configuration";
 
 static TYPE_BODY_TRACKER_CREATE_INFO_BD: Lazy<xr::StructureType> =
     Lazy::new(|| xr::StructureType::from_raw(1000385001));
@@ -122,76 +120,12 @@ type GetBodyTrackingStateBD = unsafe extern "system" fn(
     *mut BodyTrackingErrorCodeBD,
 ) -> sys::Result;
 
-#[repr(C)]
-struct MotionTrackerConnectStateBD {
-    tracker_count: i32,
-    serials: [MotionTrackerSerialBD; 6],
-}
-
-#[repr(C)]
-#[derive(Copy)]
-pub struct MotionTrackerSerialBD {
-    pub serial: [u8; 24],
-}
-impl Clone for MotionTrackerSerialBD {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-
-#[repr(transparent)]
-#[derive(Copy, Clone, Eq, PartialEq)]
-pub struct MotionTrackerConfidenceBD(i32);
-
-#[repr(C)]
-pub struct MotionTrackerLocationBD {
-    pub pose: sys::Posef,
-    pub angular_velocity: sys::Vector3f,
-    pub linear_velocity: sys::Vector3f,
-    pub angular_acceleration: sys::Vector3f,
-    pub linear_acceleration: sys::Vector3f,
-}
-
-#[repr(C)]
-pub struct MotionTrackerLocationsBD {
-    pub serial: MotionTrackerSerialBD,
-    pub local_pose: MotionTrackerLocationBD,
-    pub confidence: MotionTrackerConfidenceBD,
-    pub global_pose: MotionTrackerLocationBD,
-}
-
-type GetMotionTrackerConnectStateBD =
-    unsafe extern "system" fn(sys::Instance, *mut MotionTrackerConnectStateBD) -> sys::Result;
-
-type GetMotionTrackerLocationsBD = unsafe extern "system" fn(
-    sys::Instance,
-    sys::Time,
-    *const MotionTrackerSerialBD,
-    *mut MotionTrackerLocationsBD,
-) -> sys::Result;
-
-type SetConfigPICO = unsafe extern "system" fn(sys::Session, i32, *const c_char) -> sys::Result;
-
-#[repr(transparent)]
-#[derive(Copy, Clone, Eq, PartialEq)]
-pub struct MotionTrackerModeBD(i32);
-impl MotionTrackerModeBD {
-    pub const BODY_TRACKING: MotionTrackerModeBD = Self(0i32);
-    pub const MOTION_TRACKING: MotionTrackerModeBD = Self(1i32);
-}
-
-type GetMotionTrackerModeBD =
-    unsafe extern "system" fn(sys::Instance, *mut MotionTrackerModeBD) -> sys::Result;
-
 pub struct BodyTrackerBD {
     handle: XrBodyTrackerBD,
     session: xr::Session<AnyGraphics>,
     destroy_body_tracker: DestroyBodyTrackerBD,
     locate_body_joints: LocateBodyJointsBD,
     get_body_tracking_state: GetBodyTrackingStateBD,
-    get_motion_tracker_connect_state: GetMotionTrackerConnectStateBD,
-    get_motion_tracker_locations: GetMotionTrackerLocationsBD,
-    get_motion_tracker_mode: GetMotionTrackerModeBD,
 }
 
 impl BodyTrackerBD {
@@ -202,10 +136,7 @@ impl BodyTrackerBD {
         system: xr::SystemId,
         prompt_calibration: bool,
     ) -> xr::Result<Self> {
-        if !extra_extensions.contains(&BD_BODY_TRACKING_EXTENSION_NAME.to_owned())
-            || !extra_extensions.contains(&BD_MOTION_TRACKING_EXTENSION_NAME.to_owned())
-            || !extra_extensions.contains(&PICO_CONFIGURATION_EXTENSION_NAME.to_owned())
-        {
+        if !extra_extensions.contains(&BD_BODY_TRACKING_EXTENSION_NAME.to_owned()) {
             return Err(sys::Result::ERROR_EXTENSION_NOT_PRESENT);
         }
 
@@ -217,13 +148,6 @@ impl BodyTrackerBD {
             get_instance_proc(&session, "xrGetBodyTrackingStateBD")?;
         let destroy_body_tracker = get_instance_proc(&session, "xrDestroyBodyTrackerBD")?;
         let locate_body_joints = get_instance_proc(&session, "xrLocateBodyJointsBD")?;
-        let get_motion_tracker_connect_state =
-            get_instance_proc(&session, "xrGetMotionTrackerConnectStateBD")?;
-        let get_motion_tracker_locations =
-            get_instance_proc(&session, "xrGetMotionTrackerLocationsBD")?;
-        let set_config: SetConfigPICO = get_instance_proc(&session, "xrSetConfigPICO")?;
-        let get_motion_tracker_mode: GetMotionTrackerModeBD =
-            get_instance_proc(&session, "xrGetMotionTrackerModeBD")?;
 
         let props = super::get_props(
             &session,
@@ -245,12 +169,8 @@ impl BodyTrackerBD {
             next: ptr::null(),
             joint_set,
         };
-        unsafe {
-            //Floor height tracking origin
-            let str = CString::new("1").unwrap();
-            //Set config property for tracking origin
-            super::xr_res(set_config(session.as_raw(), 1, str.as_ptr()))?;
 
+        unsafe {
             super::xr_res(create_body_tracker(session.as_raw(), &info, &mut handle))?;
         };
 
@@ -286,73 +206,7 @@ impl BodyTrackerBD {
             destroy_body_tracker,
             locate_body_joints,
             get_body_tracking_state,
-            get_motion_tracker_connect_state,
-            get_motion_tracker_locations,
-            get_motion_tracker_mode,
         })
-    }
-
-    pub fn locate_motion_trackers(
-        &self,
-        time: xr::Time,
-    ) -> xr::Result<Vec<MotionTrackerLocationsBD>> {
-        let mut locations = Vec::with_capacity(6);
-
-        unsafe {
-            let mut tracker_mode = MotionTrackerModeBD::BODY_TRACKING;
-            super::xr_res((self.get_motion_tracker_mode)(
-                self.session.instance().as_raw(),
-                &mut tracker_mode,
-            ))?;
-
-            if tracker_mode != MotionTrackerModeBD::MOTION_TRACKING {
-                return Ok(locations);
-            }
-        }
-
-        let mut connect_state = MotionTrackerConnectStateBD {
-            tracker_count: 0,
-            serials: [MotionTrackerSerialBD { serial: [0; 24] }; 6],
-        };
-
-        unsafe {
-            super::xr_res((self.get_motion_tracker_connect_state)(
-                self.session.instance().as_raw(),
-                &mut connect_state,
-            ))?;
-
-            for i in 0..connect_state.tracker_count as usize {
-                let mut location = MotionTrackerLocationsBD {
-                    serial: MotionTrackerSerialBD { serial: [0; 24] },
-                    local_pose: MotionTrackerLocationBD {
-                        pose: xr::Posef::IDENTITY,
-                        angular_velocity: Default::default(),
-                        linear_velocity: Default::default(),
-                        angular_acceleration: Default::default(),
-                        linear_acceleration: Default::default(),
-                    },
-                    confidence: MotionTrackerConfidenceBD(0),
-                    global_pose: MotionTrackerLocationBD {
-                        pose: xr::Posef::IDENTITY,
-                        angular_velocity: Default::default(),
-                        linear_velocity: Default::default(),
-                        angular_acceleration: Default::default(),
-                        linear_acceleration: Default::default(),
-                    },
-                };
-
-                super::xr_res((self.get_motion_tracker_locations)(
-                    self.session.instance().as_raw(),
-                    time,
-                    &connect_state.serials[i],
-                    &mut location,
-                ))?;
-
-                locations.push(location);
-            }
-        }
-
-        Ok(locations)
     }
 
     pub fn locate_body_joints(
@@ -360,18 +214,6 @@ impl BodyTrackerBD {
         time: xr::Time,
         reference_space: &xr::Space,
     ) -> xr::Result<Option<Vec<BodyJointLocationBD>>> {
-        unsafe {
-            let mut tracker_mode = MotionTrackerModeBD::BODY_TRACKING;
-            super::xr_res((self.get_motion_tracker_mode)(
-                self.session.instance().as_raw(),
-                &mut tracker_mode,
-            ))?;
-
-            if tracker_mode != MotionTrackerModeBD::BODY_TRACKING {
-                return Ok(None);
-            }
-        }
-
         let mut status_code = BodyTrackingStatusCodeBD::INVALID;
         let mut error_code = BodyTrackingErrorCodeBD::INNER_EXCEPTION;
