+++ alvr/client_openxr/src/extra_extensions/mod.rs
@@ -5,6 +5,7 @@ mod eye_tracking_social;
 mod face_tracking2_fb;
 mod face_tracking_pico;
 mod facial_tracking_htc;
+mod motion_tracking_bd;
 mod multimodal_input;
 mod passthrough_fb;
 mod passthrough_htc;
@@ -16,6 +17,7 @@ pub use eye_tracking_social::*;
 pub use face_tracking2_fb::*;
 pub use face_tracking_pico::*;
 pub use facial_tracking_htc::*;
+pub use motion_tracking_bd::*;
 pub use multimodal_input::*;
 pub use passthrough_fb::*;
 pub use passthrough_htc::*;
