+++ alvr/client_openxr/src/interaction.rs
@@ -1,3 +1,4 @@
+use crate::extra_extensions::MotionTrackerBD;
 use crate::{
     extra_extensions::{
         self, BodyJointSetBD, BodyTrackerBD, BodyTrackerFB, EyeTrackerSocial, FaceTracker2FB,
@@ -72,6 +73,7 @@ pub struct FaceSources {
 pub struct BodySources {
     pub body_tracker_fb: Option<(BodyTrackerFB, usize)>,
     pub body_tracker_bd: Option<BodyTrackerBD>,
+    pub motion_tracker_bd: Option<MotionTrackerBD>,
 }
 
 #[derive(Clone)]
@@ -391,6 +393,7 @@ impl InteractionContext {
             body_sources: BodySources {
                 body_tracker_fb: None,
                 body_tracker_bd: None,
+                motion_tracker_bd: None,
             },
         }
     }
@@ -488,8 +491,8 @@ impl InteractionContext {
             "BodyTrackerFB (full set)",
             config
                 .body_tracking
-                .clone()
-                .and_then(|s| s.body_tracking_fb.into_option())
+                .as_ref()
+                .and_then(|s| s.body_tracking_fb.as_option())
                 .map(|c| c.full_body),
             || BodyTrackerFB::new(&self.xr_session, *BODY_JOINT_SET_FULL_BODY_META),
         )
@@ -517,8 +520,8 @@ impl InteractionContext {
             "BodyTrackerBD (high accuracy)",
             config
                 .body_tracking
-                .clone()
-                .and_then(|s| s.body_tracking_bd.into_option())
+                .as_ref()
+                .and_then(|s| s.body_tracking_bd.as_option())
                 .map(|c| c.high_accuracy),
             || {
                 BodyTrackerBD::new(
@@ -548,6 +551,16 @@ impl InteractionContext {
                 },
             )
         });
+
+        self.body_sources.motion_tracker_bd = create_ext_object(
+            "MotionTrackerBD (object tracking)",
+            config
+                .body_tracking
+                .as_ref()
+                .and_then(|s| s.body_tracking_bd.as_option())
+                .map(|c| c.object_tracking),
+            || MotionTrackerBD::new(self.xr_session.clone(), &self.extra_extensions),
+        );
 
         if let Some(face_tracker) = &self.face_sources.face_tracker_pico {
             face_tracker.start_face_tracking().ok();
@@ -1068,11 +1081,15 @@ pub fn get_fb_body_tracking_points(
 
 pub fn get_bd_motion_trackers(
     time: Duration,
-    body_tracker: &BodyTrackerBD,
+    motion_tracker: &MotionTrackerBD,
 ) -> Vec<(u64, DeviceMotion)> {
     let xr_time = crate::to_xr_time(time);
 
-    if let Ok(trackers) = body_tracker.locate_motion_trackers(xr_time) {
+    if let Some(trackers) = motion_tracker
+        .locate_motion_trackers(xr_time)
+        .ok()
+        .flatten()
+    {
         let mut joints = Vec::<(u64, DeviceMotion)>::with_capacity(6);
 
         let joints_ids = [
